feat: Refactor transaction handling and UI components
- Update TxHeader to use FleaThreadResponse and adjust data structure.
- Enhance TxTimeline with icons and progress bar for transaction phases.
- Modify FleaMarket transaction phases to include new states and API calls for shipping and rating.
- Introduce WaitPaymentPanel for displaying payment status and transaction details.
- Create FleaReviewModal for submitting ratings and comments with predefined templates.
- Adjust payment and address handling in EditPayment component.
- Update types for better clarity and consistency across the application.

+++ go/handler/admin.go
@@ -30,8 +30,8 @@ func (h *adminHandler) RegisterRoutes(r *mux.Router) {
 
 func (h *adminHandler) UpdateFleaConfig(w http.ResponseWriter, r *http.Request) {
 	var req struct {
-		BaseRate float64 `json:"base_rate"`
-		MaxRate  float64 `json:"max_rate"`
+		BaseRate int64 `json:"base_rate"`
+		MaxRate  int64 `json:"max_rate"`
 	}
 
 	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
